refactor: drop the redundant custom Root; use vprs default (+ vprs 3.1.0)
The per-theme route.tsx layouts scope each theme's CSS, so MmcRoot's Root was
just re-rendering the page and its collected CSS — exactly what vprs's default
Root does. Remove it and drop the `Root:` config.

The inlined Css component (the one piece MmcRoot still carried) moves to its own
src/Css.tsx; MmcHtml still needs it because the hosted Css collapses the whole
document in the no-flag `vite build --app` build. vprs 3.1.0 now fails the build
loudly on that degrade instead of shipping a broken page silently, so this
refactor is guarded: both build modes verified rendering 300 pages with full
skeletons and isolated per-theme CSS, no degrade warnings.

Bumps vite-plugin-react-server to ^3.1.0, which also carries the route.tsx
layout-CSS fix for dev:ssr that the file-router relies on.
