+++ package.json
@@ -124,7 +124,7 @@
     "typescript-plugin-css-modules": "^5.2.0",
     "vite": "^6.4.3",
     "vite-css-modules": "^1.16.0",
-    "vite-plugin-react-server": "^3.0.0",
+    "vite-plugin-react-server": "^3.1.0",
     "vitest": "^3.2.6",
     "webpack-sources": "^3.5.0"
   },
